feat: 견적요약 시트 & 예산 범위 플로팅 뷰 견적 가격 적용, 트림이 선택되었을 경우 판단
