+++ iOS/Catalog/Catalog/TrimSelection/Data/DTO/TrimRequestDTO.swift
